+++ client/src/integration-test/kotlin/net/corda/client/CordaRPCClientTest.kt
@@ -10,6 +10,9 @@ import net.corda.core.random63BitValue
 import net.corda.core.serialization.OpaqueBytes
 import net.corda.flows.CashIssueFlow
 import net.corda.flows.CashPaymentFlow
+import net.corda.node.driver.DriverBasedTest
+import net.corda.node.driver.NodeHandle
+import net.corda.node.driver.driver
 import net.corda.node.internal.Node
 import net.corda.node.services.User
 import net.corda.node.services.config.configureTestSSL