+++ tools/loadtest/src/main/kotlin/net/corda/loadtest/tests/GenerateHelpers.kt
@@ -8,7 +8,7 @@ import net.corda.core.contracts.PartyAndReference
 import net.corda.core.crypto.AnonymousParty
 import net.corda.core.crypto.Party
 import net.corda.core.serialization.OpaqueBytes
-import net.corda.flows.CashFlow
+import net.corda.flows.CashFlowCommand
 import java.util.*
 
 fun generateIssue(
@@ -16,12 +16,12 @@ fun generateIssue(
         currency: Currency,
         notary: Party,
         possibleRecipients: List<Party>
-): Generator<CashFlow.Command.IssueCash> {
+): Generator<CashFlowCommand.IssueCash> {
     return generateAmount(0, max, Generator.pure(currency)).combine(
             Generator.pure(OpaqueBytes.of(0)),
             Generator.pickOne(possibleRecipients)
     ) { amount, ref, recipient ->
-        CashFlow.Command.IssueCash(amount, ref, recipient, notary)
+        CashFlowCommand.IssueCash(amount, ref, recipient, notary)
     }
 }
 
@@ -30,19 +30,19 @@ fun generateMove(
         currency: Currency,
         issuer: AnonymousParty,
         possibleRecipients: List<Party>
-): Generator<CashFlow.Command.PayCash> {
+): Generator<CashFlowCommand.PayCash> {
     return generateAmount(1, max, Generator.pure(Issued(PartyAndReference(issuer, OpaqueBytes.of(0)), currency))).combine(
             Generator.pickOne(possibleRecipients)
     ) { amount, recipient ->
-        CashFlow.Command.PayCash(amount, recipient)
+        CashFlowCommand.PayCash(amount, recipient)
     }
 }
 
 fun generateExit(
         max: Long,
         currency: Currency
-): Generator<CashFlow.Command.ExitCash> {
+): Generator<CashFlowCommand.ExitCash> {
     return generateAmount(1, max, Generator.pure(currency)).map { amount ->
-        CashFlow.Command.ExitCash(amount, OpaqueBytes.of(0))
+        CashFlowCommand.ExitCash(amount, OpaqueBytes.of(0))
     }
 }